+++ R/Appender.R
@@ -649,10 +649,15 @@ AppenderMemory <- R6::R6Class(
 
     #' @field buffer_events A `list` of [LogEvents]. Contents of the buffer.
     buffer_events = function() {
-      ord <- get("event_order", envir = private)
-      ord <- ord - min(ord) + 1L
-      ord <- order(ord)
-      res <- get(".buffer_events", envir = private)[ord]
+      res <- get(".buffer_events", envir = private)
+
+      if (length(res)){
+        ord <- get("event_order", envir = private)
+        ord <- ord - min(ord) + 1L
+        ord <- order(ord)
+        res <- res[ord]
+      }
+
       as_event_list(
         res[!vapply(res, is.null, FALSE)]
       )
@@ -711,6 +716,7 @@ AppenderMemory <- R6::R6Class(
 
 #' Log to a memory buffer
 #'
+#' @description
 #' An Appender that Buffers LogEvents in-memory and and redirects them to other
 #' Appenders once certain conditions are met.
 #'
@@ -725,21 +731,6 @@ AppenderMemory <- R6::R6Class(
 #'   \item{`flush_on_rotate, set_flush_on_rotate`}{`TRUE` or `FALSE`: Whether
 #'     the buffer should be flushed when the Buffer is full (f.e when you close
 #'     \R). Setting this to off can have slightly negative performance impacts.}
-#' }
-#'
-#' @section Methods:
-#'
-#' \describe{
-#'   \item{`flush()`}{Manually trigger flushing}
-#'
-#'   \item{`add_appender(appender, name = NULL)`, `remove_appender(pos)`}{
-#'     Add or remove an [Appender]. Supplying a `name` is optional but
-#'     recommended. After adding an Appender with
-#'     `appender$add_appender(AppenderConsole$new(), name = "console")` you can
-#'      refer to it via `appender$appenders$console`. `remove_appender()` can
-#'      remove an Appender by position or name.
-#'    }
-#'
 #' }
 #'
 #' @export
@@ -806,6 +797,15 @@ AppenderBuffer <- R6::R6Class(
       invisible(self)
     },
 
+
+    #' @description Exactly like A [Logger], an [AppenderBuffer] can have an
+    #' arbitrary amount of Appenders attached. When the buffer is flushed, the
+    #' buffered events are dispatched to these Appenders.
+    #'
+    #' @param x single [Appender] or a `list` thereof. Appenders control the
+    #'   output of a Logger. Be aware that a Logger also inherits the Appenders
+    #'   of its ancestors (see `vignette("lgr", package = "lgr")` for more info
+    #'   about Logger inheritance).
     set_appenders = function(x){
       if (is.null(x)){
         private$.appenders <- list()
@@ -828,6 +828,15 @@ AppenderBuffer <- R6::R6Class(
     },
 
 
+    #' @description Add an Appender to the AppenderBuffer
+    #' @param appender a single [Appender]
+    #' @param name a `character` scalar. Optional but recommended.
+    #'
+    #' @description  Add or remove an [Appender]. Supplying a `name` is optional but
+    #'     recommended. After adding an Appender with
+    #'     `appender$add_appender(AppenderConsole$new(), name = "console")` you can
+    #'      refer to it via `appender$appenders$console`. `remove_appender()` can
+    #'  remove an Appender by position or name.
     add_appender = function(
       appender,
       name = NULL
@@ -842,6 +851,10 @@ AppenderBuffer <- R6::R6Class(
     },
 
 
+    #' @description remove an appender
+    #' @param pos `integer` index or `character` name of the Appender(s) to
+    #' remove
+    #'
     remove_appender = function(
       pos
     ){
@@ -971,10 +984,7 @@ AppenderFileRotating <- R6::R6Class(
 
       if (!file.exists(file))  file.create(file)
 
-      private$bq <- rotor::BackupQueueIndex$new(
-        file,
-        backup_dir = backup_dir
-      )
+      private$bq <- rotor::BackupQueueIndex$new(file)
 
       self$set_file(file)
       self$set_threshold(threshold)
@@ -1005,7 +1015,7 @@ AppenderFileRotating <- R6::R6Class(
       bq <- get("bq", private)
 
       if (force || bq$should_rotate(size = self$size)){
-        bq$push_backup()
+        bq$push()
         bq$prune()
         file.remove(self$file)
         file.create(self$file)
@@ -1025,7 +1035,7 @@ AppenderFileRotating <- R6::R6Class(
       file
     ){
       super$set_file(file)
-      private$bq$set_file(self$file)
+      private$bq$set_origin(self$file)
       self
     },
 
@@ -1066,7 +1076,7 @@ AppenderFileRotating <- R6::R6Class(
     set_backup_dir = function(
       x
     ){
-      private$bq$set_backup_dir(x)
+      private$bq$set_dir(x, create = FALSE)
       self
     },
 
@@ -1112,10 +1122,10 @@ AppenderFileRotating <- R6::R6Class(
 
     #' @field backups A `data.frame` containing information on path, file size,
     #'     etc... on the available backups of `file`.
-    backups     = function() get("bq", private)$backups,
+    backups     = function() get("bq", private)$files,
 
 
-    backup_dir  = function() get("bq", private)$backup_dir
+    backup_dir  = function() get("bq", private)$dir
   ),
 
 
@@ -1157,10 +1167,7 @@ AppenderFileRotatingTime <- R6::R6Class(
 
       if (!file.exists(file))  file.create(file)
 
-      private$bq <- rotor::BackupQueueDateTime$new(
-        file,
-        backup_dir = backup_dir
-      )
+      private$bq <- rotor::BackupQueueDateTime$new(file)
 
       self$set_file(file)
       self$set_threshold(threshold)
@@ -1196,7 +1203,7 @@ AppenderFileRotatingTime <- R6::R6Class(
           now = now
         )
       ){
-        bq$push_backup(
+        bq$push(
           overwrite = self$overwrite,
           now = now
         )
@@ -1324,10 +1331,7 @@ AppenderFileRotatingDate <- R6::R6Class(
 
       if (!file.exists(file))  file.create(file)
 
-      private$bq <- rotor::BackupQueueDate$new(
-        file,
-        backup_dir = backup_dir
-      )
+      private$bq <- rotor::BackupQueueDate$new(file)
 
       self$set_file(file)
       self$set_threshold(threshold)
@@ -1349,143 +1353,6 @@ AppenderFileRotatingDate <- R6::R6Class(
   )
 )
 
-
-# AppenderSyslog ----------------------------------------------------------
-
-#' Log to the POSIX System Log
-#'
-#' @description
-#' An Appender that writes to the syslog on supported POSIX platforms. Requires
-#' the \pkg{rsyslog} package.
-#'
-#' @seealso [LayoutFormat], [LayoutJson]
-#' @family Appenders
-#' @export
-#' @examples
-#' if (requireNamespace("rsyslog", quietly = TRUE)) {
-#'   lg <- get_logger("rsyslog/test")
-#'   lg$add_appender(AppenderSyslog$new(), "syslog")
-#'   lg$info("A test message")
-#'
-#'   if (Sys.info()[["sysname"]] == "Linux"){
-#'     system("journalctl -t 'rsyslog/test'")
-#'   }
-#'
-#'   invisible(lg$config(NULL))  # cleanup
-#' }
-AppenderSyslog <- R6::R6Class(
-  "AppenderSyslog",
-  inherit = Appender,
-  cloneable = FALSE,
-  public = list(
-    initialize = function(
-      identifier = NULL,
-      threshold = NA_integer_,
-      layout = LayoutFormat$new("%m"),
-      filters = NULL,
-      syslog_levels = c(
-        "CRITICAL" = "fatal",
-        "ERR" = "error",
-        "WARNING" = "warn",
-        "INFO" = "info",
-        "DEBUG" = "debug",
-        "DEBUG" = "trace"
-      )
-    ){
-      if (!requireNamespace("rsyslog", quietly = TRUE)) {
-        stop("The 'rsyslog' package is required for this appender.")
-      }
-      self$set_threshold(threshold)
-      self$set_layout(layout)
-      self$set_filters(filters)
-      self$set_identifier(identifier)
-      self$set_syslog_levels(syslog_levels)
-    },
-
-
-    append = function(event){
-      identifier <- get(".identifier", private)
-      if (is.null(identifier)) identifier <- event$logger
-
-      rsyslog::open_syslog(identifier = identifier)
-      rsyslog::set_syslog_mask("DEBUG")
-      on.exit(rsyslog::close_syslog())
-
-      rsyslog::syslog(
-        private$.layout$format_event(event),
-        level = private$to_syslog_levels(event$level)
-      )
-    },
-
-
-    #' @description Define conversion between lgr and syslog log levels
-    #' @param x
-    #'  * a named `character` vector mapping whose names are log
-    #'    levels as understood by [rsyslog::syslog()] and whose values are [lgr
-    #'    log levels][log_levels] (either `character` or `numeric`)
-    #'  * a `function` that takes a vector of lgr log levels as input and
-    #'    returns a `character` vector of log levels for [rsyslog::syslog()].
-    set_syslog_levels = function(x){
-      if (is.function(x)){
-        private$.syslog_levels <- x
-      } else {
-        assert(all_are_distinct(unname(x)))
-        assert(is_equal_length(x, names(x)))
-        private$.syslog_levels <- structure(
-          standardize_log_levels(unname(x)),
-          names = names(x)
-        )
-      }
-
-      self
-    },
-
-    #' @description Set a string to identify the process.
-    set_identifier = function(x){
-      assert(is.null(x) || is_scalar_character(x))
-      private$.identifier <- x
-      self
-    }
-  ),
-
-  # +- active ---------------------------------------------------------------
-  active = list(
-    destination   = function() sprintf("syslog [%s]", private$.identifier),
-
-    #' @field identifier `character` scalar. A string identifying the process;
-    #'  if `NULL` defaults to the logger name
-    identifier    = function() get(".identifier", private),
-
-    #' @field syslog_levels. Either a named `character` vector or a `function`
-    #'   mapping lgr [log_levels] to rsyslog log levels. See
-    #'   `$set_syslog_levels()`.
-    syslog_levels = function() get(".syslog_levels", private)
-  ),
-
-  private = list(
-    finalize = function(){
-      rsyslog::close_syslog()
-    },
-
-    to_syslog_levels = function(
-      levels
-    ){
-      sl <- get(".syslog_levels", private)
-      levels <- standardize_log_levels(levels)
-
-      if (is.function(sl)){
-        res <- sl(levels)
-      } else {
-        res <- names(private$.syslog_levels)[match(levels, unname(private$.syslog_levels))]
-      }
-
-      toupper(res)
-    },
-
-    .identifier = NULL,
-    .syslog_levels = NULL
-  )
-)