+++ R/utils-tests.R
@@ -1,9 +1,0 @@
-file_is_readable <- function(x){
-  suppressWarnings(
-  tryCatch({
-    readBin(x, "raw", n = 1)
-    TRUE
-  },
-  error = function(e) FALSE
-  ))
-}